renamed emu analyzer
